refactor backend test into ava
